Khaliya/requirements.txt
google-cloud-speech

google-cloud-texttospeech

google-generativeai

twilio

flask

flask-sockets

gevent

gevent-websocket

python-dotenv

requests